layout: Switch order of buttons being declared
Previously we have a 'CANCEL' & 'SUBMIT' button side by side. But the
'SUBMIT' button was declared before the 'CANCEL' button in the html
file despite the former being on the right side of the UI.

When tabbing through the form elements, this would result in us
traversing the form from right to left which we do not want.